Break out SqlTransaction into internal module
