+++ src/Database/Persist/Monad.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
@@ -82,65 +81,15 @@ import Control.Monad.Reader.Class (MonadReader (..))
 import Control.Monad.Trans.Class (MonadTrans (..))
 import Control.Monad.Trans.Resource (MonadResource)
 import Data.Pool (Pool)
-import Database.Persist.Sql (SqlBackend, SqlPersistT, runSqlConn)
-import qualified GHC.TypeLits as GHC
+import Database.Persist.Sql (SqlBackend)
 import UnliftIO.Concurrent (threadDelay)
 import UnliftIO.Exception (Exception, SomeException, catchJust, throwIO)
 import UnliftIO.Pool (withResource)
 
 import Control.Monad.IO.Rerunnable (MonadRerunnableIO)
-import Control.Monad.Trans.Rerunnable (MonadRerunnableTrans)
 import Database.Persist.Monad.Class
+import Database.Persist.Monad.Internal.SqlTransaction
 import Database.Persist.Monad.Shim
-import Database.Persist.Monad.SqlQueryRep
-
-{- SqlTransaction -}
-
-{-| The monad that tracks transaction state.
-
- Conceptually equivalent to 'Database.Persist.Sql.SqlPersistT', but restricts
- IO operations, for two reasons:
-
-   1. Forking a thread that uses the same 'SqlBackend' as the current thread
-      causes Bad Things to happen.
-   2. Transactions may need to be retried, in which case IO operations in
-      a transaction are required to be rerunnable.
-
- You shouldn't need to explicitly use this type; your functions should only
- declare the 'MonadSqlQuery' constraint.
--}
-newtype SqlTransaction m a = SqlTransaction
-  { unSqlTransaction :: SqlPersistT m a
-  }
-  deriving (Functor, Applicative, Monad, MonadFix, MonadRerunnableIO, MonadRerunnableTrans)
-
-instance
-  ( GHC.TypeError ( 'GHC.Text "Cannot run arbitrary IO actions within a transaction. If the IO action is rerunnable, use rerunnableIO")
-  , Monad m
-  ) =>
-  MonadIO (SqlTransaction m)
-  where
-  liftIO = undefined
-
-instance (MonadSqlQuery m, MonadUnliftIO m) => MonadSqlQuery (SqlTransaction m) where
-  type TransactionM (SqlTransaction m) = TransactionM m
-
-  runQueryRep = SqlTransaction . runSqlQueryRep
-
-  -- Delegate to 'm', since 'm' is in charge of starting/stopping transactions.
-  -- 'SqlTransaction' is ONLY in charge of executing queries.
-  withTransaction = SqlTransaction . withTransaction
-
-runSqlTransaction :: MonadUnliftIO m => SqlBackend -> SqlTransaction m a -> m a
-runSqlTransaction conn = (`runSqlConn` conn) . unSqlTransaction
-
--- | Errors that can occur within a SQL transaction.
-data TransactionError
-  = -- | The retry limit was reached when retrying a transaction.
-    RetryLimitExceeded
-  deriving (Show, Eq)
-
-instance Exception TransactionError
 
 {- SqlQueryT monad -}
 
@@ -230,6 +179,14 @@ mapSqlQueryT f = SqlQueryT . mapReaderT f . unSqlQueryT
 instance MonadReader r m => MonadReader r (SqlQueryT m) where
   ask = lift ask
   local = mapSqlQueryT . local
+
+-- | Errors that can occur when running a SQL transaction.
+data TransactionError
+  = -- | The retry limit was reached when retrying a transaction.
+    RetryLimitExceeded
+  deriving (Show, Eq)
+
+instance Exception TransactionError
 
 {- Running SqlQueryT -}
